Calculation for household ready
Implemented a way to calculate co2 for household. I added values to choicebox, connected button action to method, created method with checks and system prints. NOTE: all three options in choice box gives the same output which will be fixed and the output is in Integer which will be changed to Double